backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pymongo==4.5.0
motor==3.3.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.25.2
pytest==7.4.0
httpx==0.24.1
python-multipart==0.0.6
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
validators==0.22.0
nltk==3.8.1
scikit-learn==1.3.0
tqdm==4.66.1
openai==0.28.0
aiofiles==23.2.1
pytesseract==0.3.10
Pillow==10.0.0
langdetect==1.0.9
python-slugify==8.0.1